[jobs] Unify matching identities from recommendation results
This function automates the identities unify process obtaining a
list of recommendations where matching individuals can be merged.
After that, matching individuals are merged.
This job returns a list with the individuals which have been merged
and the errors generated during this process.

Individuals both for `source_uuids` and `target_uuids` are defined
by any of their valid keys or UUIDs. When the parameter
`target_uuids` is empty, the job will take all the individuals
stored in the registry, so matches will be found comparing the
identities from the individuals in `source_uuids` against all the
identities on the registry.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>