[schema] Add mutation for unify matching identities
The mutation returns a job which returns a list of identities
which have been merged with other matching identities according to
the specified criteria (a combination of email, name and username).

If the target list of individuals is empty, the job will be
applied to all the individuals in the registry.

Example for the new mutation:
```
mutation {
  unify(sourceUuids: ["<uuid1>",
                      "<uuid2>"],
        criteria: ["email", "name", "username"]){
    jobId
  }
}
```

Example for the query asking for the job result:
```
{
  job(jobId: "<job_id>"){
    jobId
    status
    jobType
    result {
      __typename
      ... on UnifyResultType {
          merged
      }
    }
    errors
  }
}
```

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>