[ui] Allow to lock individuals
This commit adds the functionality of locking or unlocking
individuals by clicking on the lock icon on the IndividualCard
component. Those mutations require the individual's uuid as an
argument, so it is added to IndividualCard as a prop and passed
from IndividualsGrid and ProfileCard.

If the mutation fails, a snackbar with the error message is
displayed on the bottom of the screen.

Signed-off-by: Eva Millán <evamillan@bitergia.com>