[ui] Add CSRF token to the Apollo petitions
This commit fixes the error produced when the app do petitions to
the Django server and the CSRF cookie is not set. To solve this,
the code forces a GET to the Django server in order to retrieve
the CSRF token, then the token is added when the Apollo server
is configured.

Signed-off-by: David Moreno <dmoreno@bitergia.com>