[ui] Remove individuals table action
Adds the 'deleteIdentity' mutation, which removes an individual.
The individual can be deleted from the actions menu in
'IndividualEntry' or from the table actions when it's selected
on 'IndividualsTable'.
A dialog appears before the individual is deleted that prompts
the user to confirm the action.
An individual can't be deleted if their profile is locked.

Signed-off-by: Eva Millán <evamillan@bitergia.com>