[ui] Edit profile information
Clicking on an individual's name, email, gender or
country on the IndividualTable opens a dialog to
edit them, unless the profile is locked.
The new information is emitted so the workspace can
be updated accordingly.
The lock and bot icons on IndividualsEntry are displayed
on hover, and can be clicked to edit them.
The individual's gender and country are added to
ExpandedIndividual.

Signed-off-by: Eva Millán <evamillan@bitergia.com>