[CI] Migrate from Travis CI to GitHub Actions
Travis CI has brought a few changes in its plans
which is causing friction to the current workflow.

This commit removes the support of Travis CI and adds
the support of GitHub Actions as the new CI service.

The muggle branch is currently excluded as it is
unstable. Any pull request or push action won't trigger
the actions workflow.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>