[api] Add method for updating enrollment dates
The method updates atomically an individual's enrollment or
set of enrollments defined by the initial date range to a new
date range.

As in the enroll method wider date ranges get merged into
the wider one, this method uses the `withdraw` and the `enroll`
API methods to apply these steps:
 * Withdraw the enrollments from the provided organization and
   an existing date range.
 * Set a new enrollment to that same organization with the new
   date range.

By default, the `force` parameter is set to `True`, as it
overwrites default dates in case a more specific date is provided
when the updated enrollment is added using the `enroll` API
method internally.

In case any of the new dates are missing, the former value for
that date will be preserved. If both new dates are missing,
the method will raise an `InvalidValueError`.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>