[schema] Add mutation for update enrollment dates
The mutation update atomically an individual's enrollment
or set of enrollments defined by the initial date range to
a new date range.

As in the enroll method wider date ranges get merged into
the wider one, this method withdraws the enrollments from
the old date range and set a new enrollment with the new
date range. By default, the `force` parameter is set to
`True`, as it overwrites default dates in case a more
specific date is provided when the updated enrollment is
added.

In case any of the new dates are missing, the former
value for that date will be preserved.

Example for the new mutation:

```
mutation {
  updateEnrollment(uuid': '3283e58(...)',
          'organization': 'Example',
          'fromDate': '2012-01-01T00:00:00+0000',
          'toDate': '2014-01-01T00:00:00+0000',
          'newFromDate': '2012-01-02T00:00:00+0000',
          'newToDate': '2013-12-31T00:00:00+0000'){
    uuid
    individual {
      mk
      enrollments {
        organization {
          name
        }
      start
      end
      }
    }
  }
}
```

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>