[ui] Save Workspace
The individuals uuids are saved on the Vuex store and local storage
when they are added to the workspace to prevent them from getting lost
when the browser is closed or reloaded.
The individuals' data is retrieved when the Dashboard is loaded by
querying getIndividualsByUuid with every uuid in the store.

Signed-off-by: Eva Millán <evamillan@bitergia.com>