[ui] Add checkbox to select all individuals
Adds a checkbox that selects or deselects all individuals
on IndividualsTable. The action buttons and search boxes
on both individuals and organizations tables are moved to
a new row to make space for it. Since locked individuals can
be selected with the checkbox, the actions performed with the
selected individuals (merge, delete, etc.) are updated to
filter the locked individuals out.

Signed-off-by: Eva Millán <evamillan@bitergia.com>