[gha] Add CI workflow action
This commit adds the github action ci workflow to the
repository. This CI workflow is triggered whenever
there is a push or pull_request event to the muggle branch.
It runs the test suite of the backend and the frontend
whenever the workflow is triggered.

The database configuration for the workflow is also updated.

https://github.com/chaoss/grimoirelab-sortinghat/issues/475.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>