[flake8] Fix flake8 errors
This commit fixes the flake8 errors present in the
muggle branch. This is necessary as a flake8 check
must be added in the github action ci workflow.

It adds a configuration to ignore some errors of a file
as it is hard to understand the logic if we format the
file. It also adds a configuration to excludes a file.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>