[ui] Add order selector to Search component
Adds an optional order selector to the Search component.
The order options passed to the component need to include
the text that will be shown on the selector and the value that
will be sent on the query.
The order can be changed from descending to ascending
by clicking a button. This will add or remove an "-" before
the order value.

Signed-off-by: Eva Millán <evamillan@bitergia.com>