[ui] Create DateInput component
Adds a DateInput component, which displays
a readonly text input that opens a date picker
when touched. The input can have the default,
filled or outlined Vuetify styles.
The component emits the date with an ISO format,
which is the format used thoughout the app,
although it is shown as YYYY-MM-DD to the user
for improved readability.

Signed-off-by: Eva Millán <evamillan@bitergia.com>