Move app files to their own directory
With this change, the files needed to deploy the SortingHat
application (WSGI, URLs, GraphQL schema) are all under
the same directory, so it's easier to distribute the app
and for web servers to use it.

Configuration files are still on their own directory because
these ones are only available as examples or for developers.
To deploy in production, users should use their own
configuration.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>