[schema] Add graphQL methods for teams
This commit adds the queries and mutations for handling Team objects in schema.py.

The `AddTeam` mutation allows users to a new team to the registry.
The `DeleteTeam` mutation allows users to delete existing teams from the registry.

The `teams` query has four filters that can be used to search through teams in
the registry. The filters are: name, term, organization and parent.

Signed-off-by: Rashmi-K-A <k.a.rashmi04@gmail.com>