[gender] Optionally disable strict name matching
Adds a 'no_strict_matching' parameter to 'recommend_gender',
which disables the name validation for well formed names.
With the strict validation, the name needs to match a 'Name
Lastname' pattern, with any number of spaces between them.
In both cases, the name needs to be at least 2 characters long.

Signed-off-by: Eva Millán <evamillan@bitergia.com>