[schema] Add `subteams` field to Team objects in GraphQL
This commit adds the `subteams` field to the Team object.
`subteams` can be resolved recursively within the query multiple times.
As a result of adding subteams, this commit modifies the resolver for
Team objects to return teams of only one level at any point of time
unless the `subteams` field is requested in query.

Signed-off-by: Rashmi-K-A <k.a.rashmi04@gmail.com>