Use prefix 'SORTINGHAT_' for configuration parameters
Predefined configuration parameters will use the prefix
'SORTINGHAT_' to identify which ones are related with the
server configuration and which ones are not.

The parameters that have been renamed are 'DEFAULT_GRAPHQL_PAGE_SIZE'
and 'AUTHENTICATION_REQUIRED'. They are now 'SORTINGHAT_API_PAGE_SIZE'
and 'SORTINGHAT_AUTHENTICATION_REQUIRED'.

Settings files were updated accordingly.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>