[server] Deploy static files with 'sortinghat-admin setup'
So far, the 'setup' command only configured the database
and the user. With this patch, it will also collect and
deploy the static files to the settings.

It is also possible to install the UI files without creating
the database. This is useful in the case of building a
docker image. This step can be done during this process
and not when the container is running. For this case,
use the flag '--only-ui'.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>