[server] Activate debugging when development mode is set
When the '--dev' flag is given, the server runs in
development mode but Django will run on debug mode
only when DEBUG setting is set to True. To avoid that
confusion, we decided that DEBUG will be always active
when the server runs in development mode.

To do this, the new environment variable 'SORTINGHAT_DEBUG'
was defined. It is activated with values 'true' or '1'.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>