[sortinghat/utils] Fix merge_date_ranges
This code fixes when we have two dates and in the first the `end`
date is the same as the `start` date of the second and the `end`
date is MAX_PERIOD_DATE. They both have the same organization.

For instance:
```
- organization: Chaoss
  start: 2000-01-01
  end: 2001-01-01
- organization: Chaoss
  start: 2001-01-01
  end: 2100-01-01
```

The current behavior only creates one enrollment (the first one)
due to the `st` is the same as `saved[1]` and the `en` is
MAX_PERIOD_DATE. Thus, the `saved[1] = min(saved[1], en)` and the
second enrollment is merged wrongly.
```
Enrollments:
  Chaoss 2000-01-01 00:00:00 2001-01-01 00:00:00
```

The fix behavior will create two enrollments because when the
`st` is not less than instead of less than or equal to `saved[1]`
it will create a new one.
```
Enrollments:
  Chaoss 2000-01-01 00:00:00 2001-01-01 00:00:00
  Chaoss 2001-01-01 00:00:00 2100-01-01 00:00:00
```

Test updated accordingly.

Signed-off-by: Quan Zhou <quan@bitergia.com>