[sortinghat/utils] Fix merge_date_ranges returns an extra date
This code fixes when the `start` and `end` dates are equal.
The current behavior returns this date range twice.

For instance:
We add the date (2015-01-01, 2015-01-01) to the method
`merge_date_ranges` and it will return the same date twice:
- (2015-01-01, 2015-01-01)
- (2015-01-01, 2015-01-01)

This happens because the first iteration we have:
- `st`: 2015-01-01
- `en`: 2015-01-01
- `saved`: (2015-01-01, 2015-01-01)

And `st` is not less than `saved[1]` then it will return `saved`
(2015-01-01, 2015-01-01) and set `saved` with the new values
(`st`, `en`). When the loop is finished it will return again
`saved` (2015-01-01, 2015-01-01).

Skip when date range (`st`, `en`) and `saved` are equal will avoid
this error.

Test updated accordingly.

Signed-off-by: Quan Zhou <quan@bitergia.com>