[model] Create Group model
Adds a 'Group' model that can be used to represent
both organizations and teams. Its objects can be
linked to other Group objects with the 'organization'
field.

This commit includes two migration files: one to apply
the changes in model.py and one to copy the data in the
'organizations' table to the new 'groups' table.

Signed-off-by: Eva Millán <evamillan@bitergia.com>