[ui] Add closable and selectable options to IndividualCard
Adds the 'closable' and 'selectable' props to the 'IndividualCard'
component, both false by default.
'closable: true' will show an button that emits a 'remove' action
on click.
'selectable: true' adds a pointer cursor and a ripple effect on click.
The workspace, which uses selectable and closable cards, is updated.

Knobs are added to the component's Storybook page to better show
all the props in action and the interaction between them.

Signed-off-by: Eva Millán <evamillan@bitergia.com>