Use Cypress for end to end testing
Replaces Puppeteer and Jest as runners with Cypress for
the end to end tests. Tests are run with 'yarn test:e2e'.
Environment variables CYPRESS_USERNAME and CYPRESS_PASSWORD
are required to log in.
The default URL can be changed with CYPRESS_BASE_URL
and the API URL with the CYPRESS_API_URL env variable.
'data-cy' attributes are added to some target elements to
comply with Cypress best practices.

Signed-off-by: Eva Millán <evamillan@bitergia.com>