Migrate SortingHat 0.7 schema on Docker startup
This commits includes a new command in sortinghat-admin command
to migrate the old database schema from SortingHat 0.7 to the
new 0.8 version.

If the database schema is old, it creates a migration of the
data, move the database to a new database with '_backup' and
restores the data using the new schema.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>