[client] Allow to not verify SSL certificates
This commit allow to perform SSL requests to a domain
without verifying its certificate with the option
`ssl_verify`.

This commit changes the class for sgqlc endpoint. It now
uses RequestsEndpoint because is easier to define a session with
the verify option to False. With HTTPEndpoint it should be necessary
to overwrite the urlopen function with a custom context with SSL
disabled.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>