[docker] Build and push Docker images
This commit builds the Docker images for the worker and server and
push the images to DockerHub. Both images are signed with sigstore
and available for for arm64 and amd64.

It is required to configure an environment 'docker-release' with the
following environment variables: DOCKERHUB_USERNAME, DOCKERHUB_TOKEN
COSIGN_PRIVATE_KEY and COSIGN_PASSWORD.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>