[Import] Add import identities backend
This commit creates a new generic backend that can be implemented
to import identities to SortingHat.

To implement the backend is important to define a NAME, a custom
'__init__' if the arguments are different from the base class, and
the 'get_individuals' method that returns the individuals using the
models from 'importer.models'.

It is designed to discover backends that implements IdentitiesImporter
at 'sortinghat.core.importer.backends'.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>