[Import] Include job and schema to import identities to SortingHat
This commit includes a new Job to import identities to SortingHat using the
data obtained from the URL using the specified backend. Custom parameters
can be specified if the backend require them.

This commit also includes the schema to create, update, delete
or query the tasks that import identities.

Update worker command to include '--with-scheduler' argument.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>