[matching] Avoid checks of matching group on unseen identities
This change allows to improve the performance by 4
when matching identities. With it, the matching
algorithm checks first if a UUID of an identity has
been already seen avoiding extra checks.

If the identity has been seen before, it'll mean
there's a match. It'll need to find on which group
of identities that particular one is. Once it is found,
the identity will be added to that matching group.

On the other hand, if it hasn't been seen yet, there's
no way to determine if the identity belongs to any
other matching group, so the algorithm won't
need to check with the other matching group if
the identity is in one of them. It will be added to
the pool for later comparisions with the remaining
unprocessed matching groups.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>