[ui] Add product tour for first time users
Redirects to the '/product-tour' route the first
time the UI is loaded on a device. The tour uses
vue-shepherd to display a series of tooltips that describe
some of the website's features and can be skipped any time.
When the tour is finished or skipped it is saved as
viewed on a cookie to avoid loading it again.

Signed-off-by: Eva Millán <evamillan@bitergia.com>