[ui] Cache the individuals queries with Apollo
Sets the fetch policy of the individuals queries to
"cache-first". When a mutation changes the data of
a cached query, Apollo uses the returned data to
update the cache automatically, so no refetching is needed.
Apollo can't update the queries if items are added or
removed from a list, so after a merge, unmerge, create
or delete, the queries need to be refetched.

Signed-off-by: Eva Millán <evamillan@bitergia.com>