Improve performance of recommender job
This commit improves the performance of the recommender job:
- It now only groups individuals that are directly related. This means
that if A = {B} and B = {A, C}, it won't create a recommendation
between A and C.
- The recommendations are now created between individuals instead of
identities. This results in fewer recommendations but is more useful if you
only need to merge individuals. If you need recommendations between
identities, set verbose to True.
- It no longer queries the database when a recommendation is created;
it follows the style "Easier to Ask Forgiveness Than Permission" (EAFP).
- The recommendations internally now return the mk of an identity, avoiding
unnecessary database queries.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>