[gitdm] Gitdm identities importer
Create a new SortingHat identities importer for Gitdm format.

This backend is configured with three parameters: a URL pointing
to the file that matches emails with organizations, an optional
URL for an aliases file that associates emails, and a flag for
email validation to verify the validity of the provided email
addresses.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>