[affiliation] Improve performance of affiliation jobs
The current algorithm was running unnecessary
calls to 'find_individual_by_uuid' when the
generic affiliation job was called. In other words,
when the list of individuals to affiliate was
all the individuals in the database.

To improve it, we have followed the same approach
we did with the matching recommendations. We get
the full list of individuals with a direct query
using Django's API.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>