[core] Add organization aliases
Adds a new table to store organizations aliases.
An organization's alias can be used to search for
that organization.
When an organization is merged into another, its
name becomes an alias of the target organization.
If a name exists as an alias, no organization can
be created with that name and viceversa.

Signed-off-by: Eva Millán <evamillan@bitergia.com>