[jobs] Define dedicated queues for tenants
So far, jobs could only run in the default RQ queue.
With this change, tenant jobs can run in dedicated
queues when needed.

To implement this feature, it was neccesary to update
the creation of the databases and queues in the
configuration file. Also, a new auto-generated config
variable named TENANTS_DEDICATED_QUEUES has been set
to determine whether a tenant runs its jobs in a
dedicated queue or not.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>