Close DB connection before job completion
RQ workers create a fork to run the function. The issue arises
when, after completing the job, the MariaDB logs display the
warning:  `Aborted connection to db. Got an error reading
communication packets`. This change ensures the database
connection is closed before the fork ends, preventing the
warning from appearing.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>