[graphene] Fix default order_by handling in queries
In newer versions of Graphene, when an input is passed as
null, it overwrites the argument’s default value as None.
This commit updates the code to fix this issue and support a
future change that will be made in a future version in which
the value will be Undefined when the filter is not defined.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>