[importer] Fix error for existing identities found during importing process
When an identity already stored in the database was found,
the identity weren't properly handled.

The code caught the exception but the database object of that
identity couldn't be found. To find an identity we need the UUID
but were using the value returned on the 'AlreadyExistsError'
which is not the UUID, it's just the values of the duplicated entry
(name, email, etc).

To fix the error, instead of using the function 'find_identity_by_uuid',
we look for it using Django's ORM to look for the tuple
(source, name, email, username) because we don't know the
UUID of the identity already inserted.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>