Include missing migrations from model changes
This commit adds two missing changes to align the
migrations schema with the models file.

- Add index on the target field in Operation.
Commit with the model change: df813e73f554e15e41ac56c8979343bc5990e0a3

- When tenant permissions were introduced, the migration and
the model diverged: the migration set the value to "user",
but the model still used "readonly". Keep the value decided in
the PR #916 in the last comment ("user").

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>